V3

+++ src/main/java/fitness_tracker/fitness/Repository/CoachRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CoachRepo extends JpaRepository<Coach,Integer> {
+public interface CoachRepo extends JpaRepository<Coach,Long> {
     List<Coach> FindByEmail(String email);
 
 }
